requirements.txt
pandas>=2.0
PyYAML>=6.0
rdflib>=7.0
pyshacl>=0.25
